Currently the Hello world is by TDD development.
I am thinking to make this app as a responsive application for the
purpose of the Image Taker.

I will give lot's of name and functionality to it in future with the
daily updates on image and etc.

I know it hard to understand here but we will undertand them one by one
in readMe file.